# Speedup backups slightly + max backup folder size config