@Glease 2024-08-27
try reenable biased locking if you are not on java 8. the builtin java thing goes through the default deflater and uses `synchronized`. neither of these are free.